Don't fork() when threads are into critical sections.
While investigating issue #504 it was found that jemalloc may produce a
deadlock in the child process when forking a parent with threads doing
allocations at the same time.

In oder to make sure that the bio.c threads (the only currently in use
within Redis) are not holding mutexes nor performing allocations this
commit modifies the code so that Redis waits for the count of pending
jobs to reach zero before calling fork().

When this happens, all the threads should be blocked into
pthread_cond_wait(), so blocked (not doing allocations) and not holding
the lock.

This commit is indirectly related to issue #504, where the deadlock is
not the actual bug reason but the indirect cause, blocking the AOF rewrite
child and allowing the AOF background rewrite buffer to go over 2GB,
resulting in the server crash.